程式碼目錄結構分層
server: 處理 router 和 伺服器相關資訊
controller: 處理業務邏輯
internal: 內部會使用到的自訂一功能套件
