+++ backend/go.mod
@@ -1,6 +1,6 @@
 module SimpleIPLocation
 
-go 1.18
+go 1.19
 
 require (
 	github.com/julienschmidt/httprouter v1.3.0
